Correcting OpenSearchVectorStore documentation for AsyncClient import problem (#252)
when AsyncOpenSearch cannot be imported, you need to make sure to
install opensearch-py[async] as documented
[here](https://github.com/opensearch-project/opensearch-py/blob/main/guides/async.md#setup).

This missing detail probably didn't arise yet, as frequently aiohttp is
installed with some other packages, which effectively solved this
problem silently (see
[here](https://github.com/opensearch-project/opensearch-py/issues/332))
but its better we install it exactly as planned by
opensearchpy-maintainers.

Open Question: are dependencies for the community-package not managed
somewhere centrally? If so, opensearch-py[async] should be added there.

Co-authored-by: Georg Wind <georg.wind@d-fine.com>
Co-authored-by: Mason Daugherty <mason@langchain.dev>